beach-cleanup-app/requirements.txt
# ==============================================================================
# Beach Cleanup App - Python Backend Requirements
# ==============================================================================
#
# This file contains all Python dependencies for the FastAPI backend server
#
# ==============================================================================
# INSTALLATION
# ==============================================================================
#
# 1. Activate the virtual environment:
#    Windows: venv\Scripts\activate
#    Mac/Linux: source venv/bin/activate
#
# 2. Install all packages:
#    pip install -r requirements.txt
#
# 3. Verify installation:
#    pip list
#
# ==============================================================================

# Web Framework
fastapi>=0.100.0              # Modern, fast web framework with automatic API docs
uvicorn[standard]>=0.23.0     # ASGI server for running FastAPI

# Authentication & Security
python-jose[cryptography]>=3.3.0  # JWT token generation and verification
passlib[bcrypt]>=1.7.4        # Password hashing with bcrypt
python-multipart>=0.0.6       # Form data and file uploads support

# Database
aiosqlite>=0.19.0             # Async SQLite database adapter

# Environment & Configuration
python-dotenv>=1.0.0          # Load environment variables from .env file

# Data Validation (comes with FastAPI)
pydantic>=2.0.0               # Data validation using Python type hints
pydantic-settings>=2.0.0      # Settings management

# Utilities
email-validator>=2.0.0        # Email validation for Pydantic models
httpx>=0.24.0                 # Async HTTP client for making requests to external APIs

# Google APIs
google-api-python-client>=2.100.0  # Google Sheets API client
google-auth>=2.23.0           # Google authentication library
